Move all files under popupextension folder
