+++ popupextension-addon/pom.xml
@@ -101,7 +101,7 @@
 							<!-- Package format version - do not change -->
 							<Vaadin-Package-Version>1</Vaadin-Package-Version>
 							<Vaadin-License-Title>${Vaadin-License-Title}</Vaadin-License-Title>
-							<Vaadin-Widgetsets>org.vaadin.addons.PopupextensionWidgetset</Vaadin-Widgetsets>
+							<Vaadin-Widgetsets>org.vaadin.addons.popupextension.PopupextensionWidgetset</Vaadin-Widgetsets>
 						</manifestEntries>
 					</archive>
 				</configuration>
