+++ popupextension-demo/src/main/java/org/vaadin/addons/popupextension/demo/test2/PopupExtensionTest2UI.java
@@ -1,6 +1,6 @@
-package org.vaadin.addons.demo.test2;
+package org.vaadin.addons.popupextension.demo.test2;
 
-import org.vaadin.addons.PopupExtension;
+import org.vaadin.addons.popupextension.PopupExtension;
 
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.ui.Button;
